move floating logic to the autosuggest hooks
